@@ -1,7 +1,8 @@
 import { Component, OnInit } from '@angular/core';
-import {FormControl, FormGroup, Validators} from '@angular/forms';
+import {UntypedFormControl, UntypedFormGroup, Validators} from '@angular/forms';
 import {Router} from '@angular/router';
-import { log } from 'console';
+import { GoogleLoginProvider, SocialAuthService, SocialUser } from '@abacritt/angularx-social-login';
+import { error, log } from 'console';
 import { Subscription } from 'rxjs';
 import { ResponseData } from 'src/app/shared/model/ResponseData';
 import { AuthService } from 'src/app/shared/services/auth/auth.service';
@@ -16,13 +17,14 @@ import { ToastService } from 'src/app/shared/services/toast-service/toast.servic
 export class LoginComponent implements OnInit {
   hidePassword = true;
   rememberMe = false;
-  public loginForm: FormGroup = {} as  FormGroup;
+  public loginForm: UntypedFormGroup = {} as  UntypedFormGroup;
   loginSubscription$:Subscription={} as Subscription 
-
+  userInfo!:SocialUser
   constructor( private router: Router,
      private authService:AuthService,
      private spinnerService:SpinnerService,
-     private toastService:ToastService
+     private toastService:ToastService,
+     private authGoogleService:SocialAuthService
      ){}
 
   ngOnInit(): void {
@@ -32,10 +34,10 @@ export class LoginComponent implements OnInit {
 
   /* This method is used to intialize the form  */
    intilalizeForm(){
-      this.loginForm = new FormGroup({
-        username: new FormControl('', [Validators.required, Validators.minLength(6)]),
-        password: new FormControl('', [Validators.required, Validators.minLength(6)]),
-        checkbox: new FormControl(false) 
+      this.loginForm = new UntypedFormGroup({
+        username: new UntypedFormControl('', [Validators.required, Validators.minLength(6)]),
+        password: new UntypedFormControl('', [Validators.required, Validators.minLength(6)]),
+        checkbox: new UntypedFormControl(false) 
       });
     }
 
@@ -45,18 +47,20 @@ export class LoginComponent implements OnInit {
       username:this.loginForm?.value?.username,
       password:this.loginForm?.value?.password
    }
-   this.spinnerService.showSpinner();
+   this.spinnerService.show();
    console.log(this.loginForm.value)
    this.loginSubscription$=this.authService.login(credential).subscribe(
     (response:ResponseData<any>)=>{
       if(response){
-        this.spinnerService.hideSpinner();
+        this.spinnerService.hide();
       console.log(response);
       let token=response.payload?.token;
-      let username=response.payload?.username
+      let username=response.payload?.username;
+      let userId=response.payload?.id
       if(token){
         localStorage.setItem("token",token)
-        localStorage.setItem("username",username)
+        localStorage.setItem("username",username);
+        localStorage.setItem("userId",userId)
         location.reload();
         this.toastService.showSuccess("Logged In SuccessFully")
         this.router.navigate(['/home'])
@@ -69,6 +73,25 @@ export class LoginComponent implements OnInit {
       
     })
 
+
+    this.authGoogleService.authState.subscribe(res=>{
+      this.userInfo=res;
+    })
+  }
+  // signInWithGoogle(): any {
+  //   console.log("hello");
+    
+  //   this.authGoogleService.signIn(GoogleLoginProvider.PROVIDER_ID).then((user)=>{
+  //      this.userInfo=user;
+  //      console.log(this.userInfo);
+       
+  //   },error=>{
+  //     console.log(error);
+      
+  //   });
+  // }
+  signOut(){
+    this.authGoogleService.signOut()
   }
 
 }
