@@ -1,12 +1,16 @@
 import { BreakpointObserver, Breakpoints } from '@angular/cdk/layout';
 import { Component, OnInit } from '@angular/core';
-import { FormControl } from '@angular/forms';
+import { UntypedFormControl } from '@angular/forms';
 import { Router } from '@angular/router';
 import { log } from 'console';
 
 import { map, Observable, shareReplay, startWith } from 'rxjs';
 import { UserProfile } from 'src/app/shared/model/UserProfile';
 import { AuthService } from 'src/app/shared/services/auth/auth.service';
+import { CartService } from 'src/app/shared/services/cart-service/cart.service';
+import { ProductService } from 'src/app/shared/services/product-service/product.service';
+import { SpinnerService } from 'src/app/shared/services/spinner-service/spinner.service';
+import { ToastService } from 'src/app/shared/services/toast-service/toast.service';
 import { TokenStorageService } from 'src/app/shared/services/token-storage/token-storage.service';
 
 @Component({
@@ -17,16 +21,133 @@ import { TokenStorageService } from 'src/app/shared/services/token-storage/token
 export class HeaderComponent implements OnInit {
   FullName:any;
   isUserLoggedIn:boolean=false;
+  cartArray:any[]=[];
+  cartItemCount=0;
+  productList:any;
+  filteredProductList:any;
+  showSearchIcon = true;
+
+  category=[
+  {
+    title: "Drone",
+    subCategory: [
+      {
+        id:1,
+        title:"Nano Drone",
+        icon:"",
+      },
+      {
+        id:2,
+        title:"Micro Drone",
+        icon:"",
+      },
+      {
+        id:3,
+        title:"Enterprise Drone",
+        icon:"",
+      },
+      {
+        id:4,
+        title:"Toy Drone",
+        icon:""
+      }
+    ]
+  },
+  {
+    title:"Services",
+    url:"assets/img/utility/drone-service.png",
+    subCategory:[
+      {
+        id:1,
+        title:"Rent a Drone",
+        icon:"",
+        products:[]
+      },
+      {
+        id:2,
+        title:"Drone Repair",
+        icon:"",
+        products:[]
+      },
+      {
+        id:3,
+        title:"Drone Training",
+        icon:"",
+        
+      },
+      {
+        id:4,
+        title:"Get Your UIN",
+        icon:"",
+        
+      },
+      {
+        id:4,
+        title:"Drone Piolet License",
+        icon:"",
+        
+      }
+    ]
+    },
+    {
+      title:"Camera",
+      url:"assets/img/utility/camera.png",
+      subCategory:[
+        {
+          id:1,
+          title:"Action Camera",
+          icon:"",
+        },
+        {
+          id:2,
+          title:"DSLR",
+          icon:""
+        },
+        {
+          id:3,
+          title:"Gimbles",
+          icon:""
+        },
+      ]
+      },
+  
+      {
+      title:"Gaming",
+      url:"assets/img/utility/gaming.png",
+      subCategory:[
+        {
+          id:1,
+          title:"XBOX",
+          icon:""
+        },
+        {
+          id:2,
+          title:"VR Headsets",
+          icon:""
+        },
+        {
+          id:3,
+          title:"Play Stations",
+          icon:""
+        },
+      ]
+      },
+  ]
   isHandset$: Observable<boolean> = this.breakpointObserver.observe(Breakpoints.Handset)
     .pipe(
       map(result => result.matches),
       shareReplay()
     );
 
+
   constructor(
     private breakpointObserver: BreakpointObserver,
     private router:Router,
-    public authService:AuthService  
+    public authService:AuthService ,
+    private productService:ProductService,
+    private toastService:ToastService ,
+    private cartService:CartService,
+    private spinnerService:SpinnerService
 ) {
   this.authService.isLoggedIn.subscribe(res=>{
     this.isUserLoggedIn=res;
@@ -37,7 +158,7 @@ export class HeaderComponent implements OnInit {
 }
   showSocialForMobile=false;
   showSocialForweb=true;
-  myControl = new FormControl();
+  myControl = new UntypedFormControl();
   categoryList = [ 
     {
       title:"Drone",
@@ -77,7 +198,7 @@ export class HeaderComponent implements OnInit {
           icon:"exit_to_app"
         }
         ];
-  options: string[] = ['Drone','Camera','DJI Products'];
+  options: string[] = [];
   filteredOptions: Observable<string[]> = new Observable<string[]>();
   selectedVal:any
   profile!: UserProfile[];
@@ -89,6 +210,10 @@ export class HeaderComponent implements OnInit {
       this.isUserLoggedIn=true
     }
     console.log( this.isUserLoggedIn+"this.isUserLoggedIn");
+
+    this.getProductOptions();
+    console.log(this.options,"this.options");
+
     this.filteredOptions = this.myControl.valueChanges.pipe(
       startWith(''),
       map(value => this._filter(value)),
@@ -110,17 +235,27 @@ export class HeaderComponent implements OnInit {
       //   code:'Change Password'
       // },
       {
+        name:'option 2',
+        code:"Account",
+      },
+      {
         name:'option 3',
         code:'Logout'
       }
     ]
 
-    
+    this.cartService.getProduct().subscribe(res=>{
+      console.log(res);
+     this.cartArray=res;
+     console.log(this.cartArray);
+     
+      this.cartItemCount=this.cartArray.length;
+    })
   }
 
   private _filter(value: string): string[] {
     const filterValue = value.toLowerCase();
-
+    console.log(this.options,"this.option23s");
     return this.options.filter(option => option.toLowerCase().includes(filterValue));
   }
   OnSearchCourse(element:string){
@@ -133,13 +268,13 @@ export class HeaderComponent implements OnInit {
     this.router.navigateByUrl(routeString)
   }
   getUserName(username:string){
-    let userNameArr=username.split('.').map((elt:string)=>{
-      return elt.charAt(0).toUpperCase() + elt.substr(1).toLowerCase()
+    let userNameArr=username?.split('.').map((elt:string)=>{
+      return elt?.charAt(0).toUpperCase() + elt?.substr(1).toLowerCase()
     })
-    return userNameArr.join(' ')
+    return userNameArr?.join(' ')
   }
   getInitial(username:string){
-    return username.charAt(0).toUpperCase();
+    return username?.charAt(0).toUpperCase();
   }
 
   generateBadgeClass(){
@@ -153,6 +288,9 @@ export class HeaderComponent implements OnInit {
     switch(this.selectedVal){
 
       case 'Change Password':
+        break;
+
+      case 'Account':
         break;
 
       case 'Logout':
@@ -172,7 +310,82 @@ export class HeaderComponent implements OnInit {
       localStorage.clear();
       this.isUserLoggedIn=false;
       this.router.navigate(["/login"])
+    }else if(key=="Account")
+    {  
+      this.router.navigate(["/account"])
     }
   }
 
+
+  opencart(){
+    this.router.navigate(["/cart"])
+  }
+
+  getProductOptions(){
+    this.productService.getAllProductList().subscribe((res:any)=>{
+      if(res){
+        this.productList=res?.payload
+        for(let key of res?.payload){    
+          this.options.push(key.name)
+        }
+        console.log("his.options"+this.options);
+        
+      }    
+   })
+  }
+
+  onOptionSelected($event:any){
+       console.log("event",$event?.option?.value);
+       for(let key of this.productList){ 
+            if(key.name==$event?.option?.value){
+              this.router.navigate(["product-details",key?.id])
+            }
+        }
+  }
+  selectedCategories: boolean[] = new Array(this.category.length).fill(false);
+
+  toggleSubcategories(index: number) {
+    this.selectedCategories[index] = !this.selectedCategories[index];
+  }
+
+  isCategorySelected(index: number): boolean {
+    return this.selectedCategories[index];
+  }
+onClickCategory(searchString:string){
+  const obj=this.searchCategoryFromList(searchString)
+  console.log("obj",obj);
+  const categoryString=obj?.category?.title;
+  const subCategoryString=obj?.subCategory[0].title;
+  this.spinnerService.show()
+   this.productService.getProductByCategorySubcategory(categoryString,subCategoryString).subscribe(
+    res=>{
+      if(res){
+        this.spinnerService.hide();
+        this.filteredProductList=res?.payload
+        const queryParams = { data: JSON.stringify(this.filteredProductList) };
+        this.router.navigate(['/filter-product'], { queryParams });
+      }
+    }
+   )
+  
+}
+  searchCategoryFromList(searchString: string): { category: any, subCategory: any[] } | null {
+    searchString = searchString.toLowerCase(); // Convert search string to lowercase
+  
+    for (const cat of this.category) {
+      const matchingSubCategory = cat.subCategory.find(subCat =>
+        subCat.title.toLowerCase().includes(searchString)
+      );
+      
+      if (matchingSubCategory) {
+        return {
+          category: cat,
+          subCategory: [matchingSubCategory]
+        };
+      }
+    }
+  
+    return null;
+  }
+  
 }
