@@ -13,20 +13,36 @@ import { AuthGuard } from './shared/guard/auth-guard/auth.guard';
 import { LoginGuard } from './shared/guard/login-guard/login.guard';
 import { CartsComponent } from './common/carts/carts.component';
 import { ProductComponent } from './common/product/product.component';
+import { CartDataResolverResolver } from './shared/resolver/cart-data-resolver.resolver';
+import { CheckoutComponent } from './common/checkout/checkout.component';
+import { PrivacyPolicyComponent } from './components/privacy-policy/privacy-policy.component';
+import { FaqComponent } from './components/faq/faq.component';
+import { BlogContentComponent } from './components/blogs/blog-content/blog-content.component';
+import { ShippingReturnPolicyComponent } from './components/shipping-return-policy/shipping-return-policy.component';
+import { FilterProductComponent } from './common/filter-product/filter-product.component';
 
 
 const routes: Routes = [
-    {path: '', component: LoginComponent,canActivate: [LoginGuard]},
-    {path: 'home', component: HomeComponent,canActivate: [AuthGuard],},
+    {path: '', component: HomeComponent},
+    {path: 'home', component: HomeComponent},
     {path: 'signup', component: SignupComponent},
     {path: 'login', component: LoginComponent,canActivate: [LoginGuard]},
     {path: 'forget-password', component: ForgetPasswordComponent},
+    {path:"blogs/:id",component:BlogContentComponent},
     {path:"blogs",component:BlogsComponent},
     {path:"about-us",component:AboutUsComponent},
     {path:"contact-us",component:ContactUsComponent},
+    {path:"filter-product",component:FilterProductComponent},
     {path:"career",component:CareersComponent},
+    {path:"privacy-policy",component:PrivacyPolicyComponent},
+    {path:"shipping-return-policy",component:ShippingReturnPolicyComponent},
+    {path:"faq",component:FaqComponent},
     {path:"cart",component:CartsComponent},
-    {path:"product-details/:id",component:ProductComponent}
+    {path:"checkout",component:CheckoutComponent},
+    {path:"content",component:BlogContentComponent},
+    {path:"product-details/:id",component:ProductComponent},
+    {path:"account",loadChildren: () => import('./modules/account-module/account/account.module').then(m => m.AccountModule)},
+    {path:"admin",loadChildren: () => import('./modules/admin/admin.module').then(m => m.AdminModule)}
 // {path: 'admin', component: DashboardComponent,
 //   canActivate: [RoleGuardService],
 //   data: {roles: 'ROLE_ADMIN'},
