@@ -1,4 +1,6 @@
 import { Component, OnInit } from '@angular/core';
+import { BlogService } from 'src/app/shared/services/blog-service/blog.service';
+import { SpinnerService } from 'src/app/shared/services/spinner-service/spinner.service';
 
 @Component({
   selector: 'app-blogs',
@@ -7,9 +9,22 @@ import { Component, OnInit } from '@angular/core';
 })
 export class BlogsComponent implements OnInit {
 
-  constructor() { }
-
+  constructor(private blogService:BlogService,private spinnerService:SpinnerService) { }
+  blogs:any[]=[];
   ngOnInit(): void {
+    this.getAllBlogs();
   }
+  
+  getAllBlogs(){
+    this.spinnerService.show();
+   this.blogService.getAllProductList().subscribe(
+    res=>{
+      if(res)
+      this.spinnerService.hide();
+        this.blogs=res?.payload;
+    }
+   )
 
+  }
+  OpenTrendingBlog(id:any){}
 }
