@@ -1,5 +1,5 @@
 import { Injectable } from '@angular/core';
-import { Observable } from 'rxjs';
+import { Observable, Subject } from 'rxjs';
 import { Product } from '../../model/product';
 import { HttpClient } from '@angular/common/http';
 import { environment } from 'src/environments/environment';
@@ -11,7 +11,6 @@ export class ProductService {
 
   constructor(private http:HttpClient) { }
 
-  
  getAllProductList():Observable<Product[]>{
     let url=environment.baseUrl +"/getAllProductList";
     return this.http.get<Product[]>(url);
@@ -20,5 +19,11 @@ export class ProductService {
     let url=environment.baseUrl +"/getProductById/" + productId;
     return this.http.get<Product>(url);
  }
+
+getProductByCategorySubcategory(productType?:string,category?:string ):Observable<any>{
+   const url = `${environment.baseUrl}/getProductsByCategoryAndSubCategory`;
+   const queryParams = `?productType=${productType}&category=${category}`;
+   return this.http.get<any>(url+queryParams);
+}
 
 }
