@@ -1,3 +1,6 @@
 export const environment = {
-  production: true
+  production: true,
+  baseUrl:"https://api.everse.co.in",
+  baseAuthLoginUrl:"https://api.everse.co.in/api/auth/login",
+  baseAuthSignupUrl:"https://api.everse.co.in/api/auth/signup"
 };
